fix: rename alert function and config
